Added assertThat___ and equalTo___ for all types understood by NSNumber. For example: assertThatInt(42, equalToInt(42))
git-svn-id: https://hamcrest.googlecode.com/svn/trunk/hamcrest-objectivec@477 2aa0e9f1-8e24-0410-adc4-1b6a3a6e65e2

+++ Source/Core/Internal/HCIntegerTypes.h
@@ -12,7 +12,9 @@
 
 #if !defined(OBJC_API_VERSION) || OBJC_API_VERSION < 2
 
-typedef long NSInteger;
-typedef unsigned long NSUInteger;
+// Define 32-bit types only. I presume that anyone in 64-bit is not building for Tiger.
+
+typedef int NSInteger;
+typedef unsigned int NSUInteger;
 
 #endif
